Scripts/py/requirements.txt
pyglm
numpy
glfw
psd_tools
imageio
watchdog
moderngl
Pillow
requests